requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
pandas
streamlit
sentencepiece  # <-- add this line
